+++ NEWS.md
@@ -2,6 +2,7 @@
 
 # DPI 2025.11
 
+-   Improved `DPI_dag()`.
 -   Fixed `DPI_curve()` for wrong (reverse) direction of DPI caused by the change of parameter order of `x` and `y` in version 2025.10.
 -   Fixed a bug caused by `dpi` parameter-object name conflict (internally) when saving `DPI()` results into a `file`.